Components

Overseer.Component — Type

The most basic Component type.

Indexing into a component with an Entity will return the data linked to that entity, indexing with a regular Int will return directly the data that is stored in the data vector at that index, i.e. generally not the storage linked to the Entity with that Int as id.

Missing docstring.

Missing docstring for PooledComponent. Check Documenter's build log for details.